refactor: region utility
