+++ MehrakCore/Services/Commands/Genshin/GenshinImageUpdaterService.cs
